+++ src/data/changelog-version.ts
@@ -1,3 +1,3 @@
 export function getLatestVersion(): { version: string; date: string } {
-  return { version: "8.80.0", date: "2026-03-07" };
+  return { version: "8.81.0", date: "2026-03-10" };
 }